initial commit of new ifconfig parser
