+++ jc/lib.py
@@ -54,6 +54,7 @@ parsers: List[str] = [
     'hosts',
     'id',
     'ifconfig',
+    'ifconfig2',
     'ini',
     'iostat',
     'iostat-s',
